Store enumeration context as soon as possible on Start
This commit addresses a race condition that exists when starting
an enumeration. In `(*enumSource).checkForData`, one case of the
select statement depends on `(*Enumeration).ctx`, which gets set
in a function called after the goroutine which depends on it has
already started, creating a race which may end in a segmentation
fault.

By setting the Context field as soon as possible when starting an
enumeration, we avoid this race entirely.