Fix memory/goroutine leaked after enumeration by filterTaskFunc
`(*Enumeration).filterTaskFunc` creates a new `stringset.Set` for use in the anonymous function it returns, but since `Close()` is never called on this set, the goroutine spawned by `stringset.New` will never return.

Any users leveraging Amass as a library would be subject to one leaked goroutine (and associated memory) after each enumeration run.

Scoping this `stringset.Set` to the lifetime of an Enumeration would allow it to be closed properly, with no impact to functionality, by passing it as an argument to `filterTaskFunc`.